style(tests): refactor tests
