+++ test/specs/commonTests/commonHelpers.js
@@ -1,0 +1,14 @@
+export default (testName, Component) => {
+  const throwError = msg => {
+    throw new Error(`${testName}: ${msg} \n  Component: ${Component && Component.name}`)
+  }
+
+  const assertRequired = (required, description) => {
+    return required || throwError(`Required ${description}, got: ${required} (${typeof required})`)
+  }
+
+  return {
+    assertRequired,
+    throwError,
+  }
+}
