+++ test/specs/commonTests/hasSubComponents.js
@@ -1,0 +1,16 @@
+import _ from 'lodash'
+
+/**
+ * Assert a component exposes other components as (static properties).
+ * @param {React.Component|Function} Component The Component.
+ * @param {React.Component[]} subComponents Array of components that should exist on Component.
+ */
+export default (Component, subComponents) => {
+  const staticValues = _.values(Component)
+
+  _.each(subComponents, subComponent => {
+    it(`has sub component ${subComponent._meta.name}`, () => {
+      staticValues.should.contain(subComponent)
+    })
+  })
+}
