+++ test/specs/commonTests/index.js
@@ -1,0 +1,11 @@
+
+export hasSubComponents from './hasSubComponents'
+export hasUIClassName from './hasUIClassName'
+
+export * from './implementsClassNameProps'
+export * from './implementsCommonProps'
+export implementsCreateMethod from './implementsCreateMethod'
+export implementsShorthandProp from './implementsShorthandProp'
+
+export isConformant from './isConformant'
+export rendersChildren from './rendersChildren'
