+++ test/specs/commonTests/isConformant.js
@@ -1,0 +1,342 @@
+import faker from 'faker'
+import _ from 'lodash'
+import React from 'react'
+import ReactDOMServer from 'react-dom/server'
+import * as semanticUIReact from 'semantic-ui-react'
+
+import { META } from 'src/lib'
+import helpers from './commonHelpers'
+import componentInfo from './componentInfo'
+import { consoleUtil, sandbox, syntheticEvent } from 'test/utils'
+
+/**
+ * Assert Component conforms to guidelines that are applicable to all components.
+ * @param {React.Component|Function} Component A component that should conform.
+ * @param {Object} [options={}]
+ * @param {Object} [options.eventTargets={}] Map of events and the child component to target.
+ * @param {Object} [options.requiredProps={}] Props required to render Component without errors or warnings.
+ */
+export default (Component, options = {}) => {
+  const { eventTargets = {}, requiredProps = {} } = options
+  const { throwError } = helpers('isConformant', Component)
+
+  // tests depend on Component constructor names, enforce them
+  if (!Component.prototype.constructor.name) {
+    throwError([
+      'Component is not a named function. This should help identify it:',
+      `static _meta = ${JSON.stringify(Component._meta, null, 2)}`,
+      `Rendered:\n${ReactDOMServer.renderToStaticMarkup(<Component />)}`,
+    ].join('\n'))
+  }
+
+  // extract componentInfo for this component
+  const {
+    _meta,
+    constructorName,
+    componentClassName,
+    filenameWithoutExt,
+  } = _.find(componentInfo, i => i.constructorName === Component.prototype.constructor.name)
+
+  // ----------------------------------------
+  // Class and file name
+  // ----------------------------------------
+  it(`constructor name matches filename "${constructorName}"`, () => {
+    constructorName.should.equal(filenameWithoutExt)
+  })
+
+  // ----------------------------------------
+  // Is exported or private
+  // ----------------------------------------
+  // detect components like: semanticUIReact.H1
+  const isTopLevelAPIProp = _.has(semanticUIReact, constructorName)
+
+  // detect sub components like: semanticUIReact.Form.Field (ie FormField component)
+  // Build a path by following _meta.parents to the root:
+  //   ['Form', 'FormField', 'FormTextArea']
+  let apiPath = []
+  let meta = _meta
+  while (meta) {
+    apiPath.unshift(meta.name)
+    meta = _.get(semanticUIReact, [meta.parent, '_meta'])
+  }
+  // Remove parent name from paths:
+  //   ['Form', 'Field', 'TextArea']
+  apiPath = apiPath.reduce((acc, next) => (
+    [...acc, next.replace(acc.join(''), '')]
+  ), [])
+
+  // find the apiPath in the semanticUIReact object
+  const isSubComponent = _.isFunction(_.get(semanticUIReact, apiPath))
+
+  if (META.isPrivate(constructorName)) {
+    it('is not exported as a component nor sub component', () => {
+      expect(isTopLevelAPIProp).to.equal(
+        false,
+        `"${constructorName}" is private (starts with  "_").` +
+        ' It cannot be exposed on the top level API'
+      )
+
+      expect(isSubComponent).to.equal(
+        false,
+        `"${constructorName}" is private (starts with "_").` +
+        ' It cannot be a static prop of another component (sub-component)'
+      )
+    })
+  } else {
+    // require all components to be exported at the top level
+    it('is exported at the top level', () => {
+      expect(isTopLevelAPIProp).to.equal(true, [
+        `"${constructorName}" must be exported at top level.`,
+        'Export it in `src/index.js`.',
+      ].join(' '))
+    })
+  }
+
+  if (_meta.parent) {
+    it('is a static component on its parent', () => {
+      expect(isSubComponent).to.equal(
+        true,
+        `\`${constructorName}\` is a child component (has a _meta.parent).` +
+        ` It must be a static prop of its parent \`${_meta.parent}\``
+      )
+    })
+  }
+
+  // ----------------------------------------
+  // Props
+  // ----------------------------------------
+  it('spreads user props', () => {
+    // JSX does not render custom html attributes so we prefix them with data-*.
+    // https://facebook.github.io/react/docs/jsx-gotchas.html#custom-html-attributes
+    const props = {
+      [`data-${_.kebabCase(faker.hacker.noun())}`]: faker.hacker.verb(),
+    }
+
+    // descendants() accepts an enzyme <selector>
+    // props should be spread on some descendant
+    // we find the descendant with spread props via a matching props object selector
+    // we do not test Component for props, of course they exist as we are spreading them
+    shallow(<Component {...requiredProps} {...props} />)
+      .should.have.descendants(props)
+  })
+
+  describe('"as" prop (common)', () => {
+    it('renders the component as HTML tags or passes "as" to the next component', () => {
+      // silence element nesting warnings
+      consoleUtil.disableOnce()
+
+      const tags = ['a', 'em', 'div', 'h1', 'h2', 'h3', 'h4', 'h5', 'h6', 'i', 'p', 'span', 'strong']
+      try {
+        tags.forEach((tag) => {
+          shallow(<Component as={tag} />)
+            .should.have.tagName(tag)
+        })
+      } catch (err) {
+        tags.forEach((tag) => {
+          const wrapper = shallow(<Component as={tag} />)
+          wrapper.type().should.not.equal(Component)
+          wrapper.should.have.prop('as', tag)
+        })
+      }
+    })
+
+    it('renders as a functional component or passes "as" to the next component', () => {
+      const MyComponent = () => null
+
+      try {
+        shallow(<Component as={MyComponent} />)
+          .type()
+          .should.equal(MyComponent)
+      } catch (err) {
+        const wrapper = shallow(<Component as={MyComponent} />)
+        wrapper.type().should.not.equal(Component)
+        wrapper.should.have.prop('as', MyComponent)
+      }
+    })
+
+    it('renders as a ReactClass or passes "as" to the next component', () => {
+      class MyComponent extends React.Component {
+        render() {
+          return <div data-my-react-class />
+        }
+      }
+
+      try {
+        shallow(<Component as={MyComponent} />)
+          .type()
+          .should.equal(MyComponent)
+      } catch (err) {
+        const wrapper = shallow(<Component as={MyComponent} />)
+        wrapper.type().should.not.equal(Component)
+        wrapper.should.have.prop('as', MyComponent)
+      }
+    })
+
+    it('passes extra props to the component it is renders as', () => {
+      const MyComponent = () => null
+
+      shallow(<Component as={MyComponent} data-extra-prop='foo' />)
+        .should.have.descendants('[data-extra-prop="foo"]')
+    })
+  })
+
+  describe('handles props', () => {
+    it('defines handled props in Component.handledProps', () => {
+      Component.should.have.any.keys('handledProps')
+      Component.handledProps.should.be.an('array')
+    })
+
+    it('Component.handledProps includes all handled props', () => {
+      const computedProps = _.union(
+        Component.autoControlledProps,
+        _.keys(Component.defaultProps),
+        _.keys(Component.propTypes),
+      )
+      const expectedProps = _.uniq(computedProps).sort()
+
+      Component.handledProps.should.to.deep.equal(expectedProps,
+        'It seems that not all props were defined in Component.handledProps, you need to check that they are equal ' +
+        'to the union of Component.autoControlledProps and keys of Component.defaultProps and Component.propTypes'
+      )
+    })
+  })
+
+  // ----------------------------------------
+  // Events
+  // ----------------------------------------
+  it('handles events transparently', () => {
+    // Events should be handled transparently, working just as they would in vanilla React.
+    // Example, both of these handler()s should be called with the same event:
+    //
+    //   <Button onClick={handler} />
+    //   <button onClick={handler} />
+    //
+    // This test catches the case where a developer forgot to call the event prop
+    // after handling it internally. It also catch cases where the synthetic event was not passed back.
+    _.each(syntheticEvent.types, ({ eventShape, listeners }) => {
+      _.each(listeners, listenerName => {
+        // onKeyDown => keyDown
+        const eventName = _.camelCase(listenerName.replace('on', ''))
+
+        // onKeyDown => handleKeyDown
+        const handlerName = _.camelCase(listenerName.replace('on', 'handle'))
+
+        const handlerSpy = sandbox.spy()
+        const props = {
+          ...requiredProps,
+          [listenerName]: handlerSpy,
+          'data-simulate-event-here': true,
+        }
+
+        const wrapper = shallow(<Component {...props} />)
+
+        const eventTarget = eventTargets[listenerName]
+          ? wrapper.find(eventTargets[listenerName])
+          : wrapper.find('[data-simulate-event-here]')
+
+        eventTarget.simulate(eventName, eventShape)
+
+        // give event listeners opportunity to cleanup
+        if (wrapper.instance() && wrapper.instance().componentWillUnmount) {
+          wrapper.instance().componentWillUnmount()
+        }
+
+        // <Dropdown onBlur={handleBlur} />
+        //                   ^ was not called once on "blur"
+        const leftPad = ' '.repeat(constructorName.length + listenerName.length + 3)
+
+        handlerSpy.calledOnce.should.equal(true,
+          `<${constructorName} ${listenerName}={${handlerName}} />\n` +
+          `${leftPad} ^ was not called once on "${eventName}".` +
+          'You may need to hoist your event handlers up to the root element.\n'
+        )
+
+        let expectedArgs = [eventShape]
+        let errorMessage = 'was not called with (event)'
+
+        if (_.has(Component.propTypes, listenerName)) {
+          expectedArgs = [eventShape, props]
+          errorMessage = 'was not called with (event, data)'
+        }
+
+        // Components should return the event first, then any data
+        handlerSpy.calledWithMatch(...expectedArgs).should.equal(true,
+          `<${constructorName} ${listenerName}={${handlerName}} />\n` +
+          `${leftPad} ^ ${errorMessage}\n` +
+          'It was called with args:\n' +
+          JSON.stringify(handlerSpy.args, null, 2)
+        )
+      })
+    })
+  })
+
+  // ----------------------------------------
+  // Defines _meta
+  // ----------------------------------------
+  describe('_meta', () => {
+    it('is a static object prop', () => {
+      expect(_meta).to.be.an('object')
+    })
+
+    describe('name', () => {
+      it('is defined', () => {
+        expect(_meta).to.have.any.keys('name')
+      })
+      it('matches the filename', () => {
+        expect(_meta.name).to.equal(filenameWithoutExt)
+      })
+    })
+    if (_.has(_meta, 'parent')) {
+      describe('parent', () => {
+        it('matches some component name', () => {
+          expect(_.map(semanticUIReact, c => c.prototype.constructor.name)).to.contain(_meta.parent)
+        })
+      })
+    }
+    describe('type', () => {
+      it('is defined', () => {
+        expect(_meta).to.have.any.keys('type')
+      })
+      it('is a META.TYPES value', () => {
+        expect(_.values(META.TYPES)).to.contain(_meta.type)
+      })
+    })
+  })
+
+  // ----------------------------------------
+  // Handles className
+  // ----------------------------------------
+  describe('className (common)', () => {
+    it(`has the Semantic UI className "${componentClassName}"`, () => {
+      const wrapper = render(<Component {...requiredProps} />)
+      // don't test components with no className at all (i.e. MessageItem)
+      if (wrapper.prop('className')) {
+        wrapper.should.have.className(componentClassName)
+      }
+    })
+
+    it("applies user's className to root component", () => {
+      const classes = faker.hacker.phrase()
+      shallow(<Component {...requiredProps} className={classes} />)
+        .should.have.className(classes)
+    })
+
+    it("user's className does not override the default classes", () => {
+      const defaultClasses = shallow(<Component {...requiredProps} />)
+        .prop('className')
+
+      if (!defaultClasses) return
+
+      const userClasses = faker.hacker.verb()
+      const mixedClasses = shallow(<Component {...requiredProps} className={userClasses} />)
+        .prop('className')
+
+      defaultClasses.split(' ').forEach((defaultClass) => {
+        mixedClasses.should.include(defaultClass, [
+          'Make sure you are using the `getUnhandledProps` util to spread the `rest` props.',
+          'This may also be of help: https://facebook.github.io/react/docs/transferring-props.html.',
+        ].join(' '))
+      })
+    })
+  })
+}
